java/travis: run unit tests w/ missing credentials
In our travis.yml, I ignore the failure to extract the secret
credentials file. Extraction fails for all external pull requests: we
don't want Travis to just stop there, we want to run as many tests as we
are able.

I split the Java tests into separate unit tests and integration (system)
tests. This way, we can skip the integration tests for external pull
requests, but still run everything else (including checkstyle).

I noticed that I had a bunch of unused imports in the test files, so I
made checkstyle run on the test files, as well. The way the integration
test runner knows that test is an integration test is with the IT
suffix. Since this violates the checkstyle rules, I add an annotation to
ignore just this rule in those classes.